apps/contactus/api_v1/serializers.py
from rest_framework import serializers
from apps.contactus.models import ContactUs


class ContactUsSerializer(serializers.ModelSerializer):
    class Meta:
        model = ContactUs
        fields = "__all__"
