requirements.txt
Django==4.0.3
djangorestframework==3.13.1
django-environ==0.8.1
psycopg2==2.9.3
django-cors-headers==3.11.0
requests==2.27.1
razorpay==1.3.0